+++ packages/dsw-mailer/CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add 2FA verification email
+
 
 ## [3.20.2]